requirements.txt
Django==3.1.5
djangorestframework==3.13.1
psycopg2==2.9.3
django-environ==0.8.1
celery==5.2.3
gevent==21.12.0
uvicorn==0.17.6
channels==3.0.4
django-filter==21.1
drf-yasg==1.20.0
django-celery-beat==2.2.1
django-debug-toolbar==3.2
redis==3.5.3